Fix for the Channel row of the property grid not having its limits updated when the protocol automatically switches to Twinkly
Repro:
1) Add a new Ethernet controller
2) Set Vendor to Twinkly
3) Set Model to Cluster (notice this also updates the Protocol to 'Twinkly')
4) Untick Auto Size
5) Change 'Channels' to 1200.
Notice that when you defocus the Channels property (eg by selecting another property), it will revert back to 510.

Workaround: After step 3, switch to a different controller and back, to rebuild the property grid.

This happened because Twinkly controllers only have one protocol available, and this protocol is auto-selected (by ControllerEthernet::UpdateProperties), but in doing so by-passes the normal handler that would rebuild the property grid in response.
I've fixed this by reusing the same logic that triggers in the property grid event handler, which rebuilds the property grid.